CMakeList modified, plus libraries added, to compile to Release folder, on Windows (MSVC 2019). Shader code slightly modified to account for non variable size array issue.

+++ src/objparser.cpp
@@ -29,7 +29,7 @@ bool ObjParser::parseFile()
   auto startParsing = std::chrono::system_clock::now();
 
   std::string current;
-  static u_int8_t mainCounter = 0;
+  static int mainCounter = 0;
 
   while (std::getline(m_objFile, current))
     {
